Update Programmers Code (Java)

+++ Programmers/코딩테스트 입문/Java/Lv.0/정리완료/[Lv.0] 머쓱이보다 키 큰 사람/[Lv.0] 머쓱이보다 키 큰 사람_01.java
@@ -5,6 +5,6 @@ class Solution {
         // 배열 스트림 생성 -> 원소 추출 -> 개수 계산
         // filter(): 조건에 해당하는 원소를 추출하여 반환
         // count(): 개수 반환
-        return Arrays.stream(array).filter(i -> i > height).count();
+        return (int) Arrays.stream(array).filter(i -> i > height).count();
     }
 }
